fix: separated fixture to fix quality issues
