+++ dependencies_health/check_django_dependencies_compatibility.py
@@ -4,7 +4,8 @@ Contains check that reads/parses dependencies of a repo
 
 import logging
 
-from repo_health.check_django_dependencies_compatibility import django_dependency_sheet_fixture, set_django_packages # pylint: disable=unused-import
+from repo_health.check_django_dependencies_compatibility import set_django_packages
+from repo_health.fixtures.config_files import django_dependency_sheet_fixture  # pylint: disable=unused-import
 
 logger = logging.getLogger(__name__)
 
@@ -18,6 +19,5 @@ def check_django_dependencies_status(repo_path, all_results, django_dependency_s
     all_results = set_django_packages(
         repo_path,
         all_results,
-        django_dependency_sheet_fixture,
-        MODULE_DICT_KEY
+        django_dependency_sheet_fixture
     )
