+++ dependencies_health/check_setup_py.py
@@ -5,8 +5,9 @@ And gathers info
 
 import re
 
-from repo_health.check_setup_py import (fixture_python_version, fixture_setup_cfg, fixture_setup_py, set_pypi_name,
-                                        set_repo_url) # pylint: disable=unused-import
+from repo_health.check_setup_py import set_pypi_name, set_repo_url
+from repo_health.fixtures.config_files import (fixture_python_version,  # pylint: disable=unused-import
+                                               fixture_setup_cfg, fixture_setup_py)
 
 module_dict_key = "setup_py"
