+++ repo_health/fixtures/__init__.py
@@ -1,0 +1,3 @@
+"""
+Contains the fixtures using in repo and dependencies health checks
+"""
